LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-viewport-units-expected.txt



FAIL 10vw is 40px assert_equals: expected "40px" but got ""
FAIL 10vh is 20px assert_equals: expected "20px" but got ""
FAIL 10vi is 40px assert_equals: expected "40px" but got ""
FAIL 10vb is 20px assert_equals: expected "20px" but got ""
FAIL 10vmin is 20px assert_equals: expected "20px" but got ""
FAIL 10vmax is 40px assert_equals: expected "40px" but got ""
FAIL 10svw is 40px assert_equals: expected "40px" but got ""
FAIL 10svh is 20px assert_equals: expected "20px" but got ""
FAIL 10svi is 40px assert_equals: expected "40px" but got ""
FAIL 10svb is 20px assert_equals: expected "20px" but got ""
FAIL 10svmin is 20px assert_equals: expected "20px" but got ""
FAIL 10svmax is 40px assert_equals: expected "40px" but got ""
FAIL 10lvw is 40px assert_equals: expected "40px" but got ""
FAIL 10lvh is 20px assert_equals: expected "20px" but got ""
FAIL 10lvi is 40px assert_equals: expected "40px" but got ""
FAIL 10lvb is 20px assert_equals: expected "20px" but got ""
FAIL 10lvmin is 20px assert_equals: expected "20px" but got ""
FAIL 10lvmax is 40px assert_equals: expected "40px" but got ""
FAIL 10dvw is 40px assert_equals: expected "40px" but got ""
FAIL 10dvh is 20px assert_equals: expected "20px" but got ""
FAIL 10dvi is 40px assert_equals: expected "40px" but got ""
FAIL 10dvb is 20px assert_equals: expected "20px" but got ""
FAIL 10dvmin is 20px assert_equals: expected "20px" but got ""
FAIL 10dvmax is 40px assert_equals: expected "40px" but got ""
